requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
fastapi==0.121.3
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
torch==2.9.1+cpu
torchvision==0.24.1+cpu
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
utility==1.0
uvicorn==0.38.0
